+++ test/dummy/app/controllers/pdf_controller.rb
@@ -1,9 +1,0 @@
-class PdfController < ApplicationController
-  
-  def render_pdf
-   @items = [{:name=> "Hello"},{:name=> "World"}]
-   render "render_pdf.pdf.prawn"
-  end
-
- 
-end