+++ test/dummy/config/initializers/prawn-rails.rb
@@ -1,4 +1,0 @@
-PrawnRails.config do |config|
-  config.page_layout = :landscape
-end
-