+++ test/dummy_app/app/controllers/reports_controller.rb
@@ -1,0 +1,7 @@
+class ReportsController < ApplicationController
+  
+  def sample
+    @items = [{:name=> "Hello"},{:name=> "World"}]
+  end
+ 
+end