+++ test/prawn_rails_test.rb
@@ -1,11 +1,17 @@
+### FOR GENERIC NON-RAILS TESTS
+
 require 'test_helper'
-require "prawn-rails/extension"
 
 class PrawnRailsTest < ActiveSupport::TestCase
-  include PrawnRails::Extension
 
+  # TODO remove this test
+  include PrawnRails::Extension
   test "html_strip" do
     assert_nil html_strip(nil)
     assert_not_nil html_strip("")
+  end
+
+  test "config" do
+    assert true
   end
 end